database events

+++ Classes/Hooks/TCEmainHook.php
@@ -11,9 +11,13 @@ namespace TRAW\EventDispatch\Hooks;
  * The TYPO3 project - inspiring people to share!
  */
 
-use TRAW\EventDispatch\Events\Database\AfterDatabaseOperationEvent;
 use TRAW\EventDispatch\Events\Cache\ClearCacheEvent;
+use TRAW\EventDispatch\Events\Database\AfterDatabaseOperationEvent;
 use TRAW\EventDispatch\Events\Database\DeleteRecordEvent;
+use TRAW\EventDispatch\Events\Database\MoveRecordEvent;
+use TRAW\EventDispatch\Events\Database\PostProcessEvent;
+use TRAW\EventDispatch\Events\Database\PostProcessFieldArrayEvent;
+use TRAW\EventDispatch\Events\Database\PreProcessFieldArrayEvent;
 use TYPO3\CMS\Core\DataHandling\DataHandler;
 
 
@@ -32,7 +36,50 @@ class TCEmainHook extends AbstractHook
      */
     public function processCmdmap_deleteAction(string $table, int $id, array $recordToDelete, ?bool $recordWasDeleted, DataHandler &$pObj)
     {
-        $this->triggerEvent(new DeleteRecordEvent($table, $id, $recordToDelete, $recordWasDeleted, $pObj));
+        if ($this->settings->getDeleteRecord()) {
+            $this->triggerEvent(new DeleteRecordEvent($table, $id, $recordToDelete, $recordWasDeleted, $pObj));
+        }
+    }
+
+    /**
+     * @param array $incomingFieldArray
+     * @param string $table
+     * @param $id
+     * @param DataHandler $dataHandler
+     */
+    public function processDatamap_preProcessFieldArray(array $incomingFieldArray, string $table, $id, DataHandler $dataHandler)
+    {
+        if ($this->settings->getPreProcessFieldArray()) {
+            $this->triggerEvent(new PreProcessFieldArrayEvent($incomingFieldArray, $table, $id, $dataHandler));
+        }
+    }
+
+    /**
+     * @param string $status
+     * @param string $table
+     * @param $id
+     * @param array $fieldArray
+     * @param DataHandler $dataHandler
+     */
+    public function processDatamap_postProcessFieldArray(string $status, string $table, $id, array $fieldArray, DataHandler $dataHandler)
+    {
+        if ($this->settings->getPostProcessFieldArray()) {
+            $this->triggerEvent(new PostProcessFieldArrayEvent($status, $table, $id, $fieldArray, $dataHandler));
+        }
+    }
+
+    /**
+     * @param string $command
+     * @param string $table
+     * @param $recordId
+     * @param $commandValue
+     * @param DataHandler $dataHandler
+     */
+    public function processCmdmap_postProcess(string $command, string $table, $recordId, $commandValue, DataHandler $dataHandler)
+    {
+        if ($this->settings->getPostProcess()) {
+            $this->triggerEvent(new PostProcessEvent($command, $table, $recordId, $commandValue, $dataHandler));
+        }
     }
 
     /**
@@ -44,7 +91,40 @@ class TCEmainHook extends AbstractHook
      */
     public function processDatamap_afterDatabaseOperations($status, $table, $id, $fieldArray, DataHandler &$pObj)
     {
-        $this->triggerEvent(new AfterDatabaseOperationEvent($status, $table, $id, $fieldArray, $pObj));
+        if ($this->settings->getAfterAllDatabaseOperations()) {
+            $this->triggerEvent(new AfterDatabaseOperationEvent($status, $table, $id, $fieldArray, $pObj));
+        }
+    }
+
+    /**
+     * @param $table
+     * @param $recordId
+     * @param $destinationPid
+     * @param array $movedRecord
+     * @param array $updatedFields
+     * @param DataHandler $dataHandler
+     */
+    public function moveRecord_firstElementPostProcess($table, $recordId, $destinationPid, array $movedRecord, array $updatedFields, DataHandler $dataHandler)
+    {
+        if ($this->settings->getMoveRecord()) {
+            $this->triggerEvent(new MoveRecordEvent($table, $recordId, $destinationPid, null, $movedRecord, $updatedFields, $dataHandler));
+        }
+    }
+
+    /**
+     * @param $table
+     * @param $recordId
+     * @param $destinationPid
+     * @param $originalDestinationPid
+     * @param array $movedRecord
+     * @param array $updatedFields
+     * @param DataHandler $dataHandler
+     */
+    public function moveRecord_afterAnotherElementPostProcess($table, $recordId, $destinationPid, $originalDestinationPid, array $movedRecord, array $updatedFields, DataHandler $dataHandler)
+    {
+        if ($this->settings->getMoveRecord()) {
+            $this->triggerEvent(new MoveRecordEvent($table, $recordId, $destinationPid, $originalDestinationPid, $movedRecord, $updatedFields, $dataHandler));
+        }
     }
 
     /**
@@ -53,6 +133,8 @@ class TCEmainHook extends AbstractHook
      */
     public function clearCachePostProc(array &$params, DataHandler &$pObj)
     {
-        $this->triggerEvent(new ClearCacheEvent($params, $pObj));
+        if ($this->settings->getClearCache()) {
+            $this->triggerEvent(new ClearCacheEvent($params, $pObj));
+        }
     }
 }
